pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = ""
authors = [
    {name = "funnymilk",email = "andre.tum91@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic[email] (>=2.11.10,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "confluent-kafka (>=2.12.1,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
[dependency-groups]
dev = [
    "pytest-cov (>=7.0.0,<8.0.0)"
]
